requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.1
Flask-Limiter==3.8.0
gunicorn==22.0.0
prometheus-client==0.21.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
cryptography==42.0.8
requests==2.32.3
APScheduler==3.10.4
numpy==1.26.4
python-dateutil==2.9.0.post0
pytest==8.3.2
pytest-flask==1.3.0
